% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boltzmann.test.R
\name{boltzmann.test}
\alias{boltzmann.test}
\alias{boltzmann.test.default}
\alias{boltzmann.test.formula}
\title{Perform a Boltzmann Test for}
\usage{
boltzmann.test(x, ...)

\method{boltzmann.test}{default}(x, y = NULL, mu = 0, paired = FALSE, ...)

\method{boltzmann.test}{formula}(formula, data, nu = 0, ...)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values}

\item{...}{further arguments passed to the different methods}

\item{y}{an optional (non-empty) numeric vector of data values}

\item{mu}{a single number indicating the true value of the mean or the
difference in means when performing a two sample test}

\item{paired}{a logical indicating whether to perform a paired test (default
FALSE)}

\item{formula}{a formula either of the form `lhs` ~ `rhs`, where `lhs` is
either a numeric or factor variable or combinations thereof (see details) and
`rhs` is one or more factor variable used to define the groups, or of the
form ~ `rhs` where `rhs` contains one or more numeric or factor variable
(see details)}

\item{data}{a data.frame containing the variables used in `formula`}

\item{nu}{a single number or a numeric vector specifying the values of the
expectations}
}
\value{
a list with class `boltzmannTestResult` containing the follow
components
\describe{
 \item{statistic}{the value of \eqn{\chi^2} statistic}
 \item{iDivergence}{the value of the I-divergence, a.k.a. Kullback-Leibler
 divergence of the hypothesis distribution or ambient alternative
 distribution to the empirical distribution}
 \item{degreesOfFreedom}{ the degrees of freedom used for computing the
 p-value via the \eqn{\chi^2} distribution with `degreesOfFreedom`}
 \item{sampleSize}{the sample size}
 \item{pValue}{the p-value}
 \item{alternativeDistribution}{the empirical distribution of outcomes}
 \item{alternativeExpectations}{the values of empirical expectations}
 \item{hypothesisDistribution}{the hypothesis distribution of outcomes}
 \item{hypothesisExpectations}{the values of hypothesized expectations}
 \item{testedExpectations}{an integer vector with the indices of the tested
   expectations}
 \item{dataName}{name of the input data}
 \item{coefficientMatrix}{an numeric matrix with the coefficient matrix}
 \item{method}{a string signifying the method}
 \item{ambientDistribution}{the ambient alternative distribution of outcomes
   if nested hypothesis testing is performed otherwise NULL}
 \item{ambientExpectations}{the values of the ambient expectations if nested
   hypothesis testing is performed otherwise NULL}
}
}
\description{
Performs a one-, two-, and multi-groups Boltzmann Test for the
the mean (one-group), the (paired) difference of the means (two-groups),
means and/or prevalences of multiple observables (class formula), means
and/or prevalences of multiple observables and multi-groups (class formula)
}
\details{
The `formula` API allows to specify multi-group comparisons for one or more
target variable specified at `lhs`. If a target variable is numeric then the
difference of means of all groups except the first versus the mean of the
first group is compared. if a target variable is a factor then the difference
of the prevalence of the second (third, etc) level of that level of all
groups except the first versus the first group is compared.If only a `rhs`
is given, a multivariate comparison to the given `nu` is performed. If the
target is numeric the mean is calculated. If the target is a factor the
prevalence is calculated.
}
\examples{
## examples using the NHANES data
data(nhanes)
## testing whether females are on average the same height as males
with(
  nhanes,
  boltzmann.test(
    x = BMXHT[RIAGENDR == "female"],
    y = BMXHT[RIAGENDR == "male"]
   )
)

## testing whether females weigh on average 10 kg less than males
boltzmann.test(BMXWT ~ RIAGENDR, data = nhanes, nu = -10)

## testing whether females have the same average weight and height as males
boltzmann.test(cbind(BMXWT, BMXHT) ~ RIAGENDR, data = nhanes)

data(kidneyStones)
## testing whether the treatment success is the same in treatment A versus B
boltzmann.test(success ~ treatment, data = kidneyStones)

## testing whether the treatment success is the same in treatment A versus B
## and small and large stone sizes (four groups)
boltzmann.test(success ~ treatment + stoneSize, data = kidneyStones)

}
