% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boltzmannTest.R
\name{boltzmannTest.entities_tibble}
\alias{boltzmannTest.entities_tibble}
\title{Compute the Boltzmann Test}
\usage{
\method{boltzmannTest}{entities_tibble}(
  entities,
  G,
  eta,
  testedMoments,
  nested = NULL,
  strict = FALSE,
  maxit = 10000L,
  tolerance = .Machine$double.eps
)
}
\description{
Compute the Boltzmann Test
}
