% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boltzmannTest.R
\name{boltzmannTest.outcomes_tibble}
\alias{boltzmannTest.outcomes_tibble}
\title{Boltzmann Test}
\usage{
\method{boltzmannTest}{outcomes_tibble}(
  outcomes,
  G,
  eta,
  testedMoments,
  nested = NULL,
  strict = FALSE,
  maxit = 10000L,
  tolerance = .Machine$double.eps
)
}
\description{
Boltzmann Test
}
