% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boltzmannTest-package.R
\docType{package}
\name{boltzmannTest}
\alias{boltzmannTest-package}
\alias{boltzmannTest}
\title{boltzmannTest}
\description{
blbla
}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
