% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boltzmannTestResult.R
\name{boltzmannTestResult}
\alias{boltzmannTestResult}
\title{Build a `boltzmannTestResult` class object}
\usage{
boltzmannTestResult(
  statistic,
  iDivergence,
  degreesOfFreedom,
  sampleSize,
  pValue,
  alternativeDistribution,
  alternativeExpectations,
  hypothesisDistribution,
  hypothesisExpectations,
  testedExpectations,
  dataName,
  coefficientMatrix,
  method = "Boltzmann Test",
  ambientDistribution = NULL,
  ambientExpectations = NULL
)
}
\arguments{
\item{statistic}{a single numeric value corresponding to the value of the (test) statistic}

\item{iDivergence}{a single numeric value corresponding to the value of the I-Divergence
of the hypothesis distribution or outer alternative distribution to the (inner) alternative
distribution}

\item{degreesOfFreedom}{a single integer value for the degrees of freedom of the Chi^2 distribution.
Gives the number of tested generalized Expectations and must be > 0}

\item{sampleSize}{a single integer value for the sample size. Must be > 1}

\item{pValue}{a single numeric value between 0 and 1 for the p-value}

\item{alternativeDistribution}{a numeric vector for the alternative distribution (usually equal
to the empirical distribution)}

\item{alternativeExpectations}{a numeric vector for the values of the generalized moments
of the alternative distribution}

\item{hypothesisDistribution}{a numeric vector for the hypothesis distribution}

\item{hypothesisExpectations}{a numeric vector for the values of the generalized moments
of the hypothesis distribution}

\item{testedExpectations}{integer vector with the indices of the tested moments}

\item{dataName}{string with the name of the data}

\item{coefficientMatrix}{a numeric matrix holding the coefficient matrix to calculate
the generalized moments. The number of rows must equal the number of generalized moments and the number
of columns the number of entities.}

\item{ambientDistribution}{a numeric vector for the ambient alternative distribution}

\item{ambientExpectations}{a numeric vector for the values of the generalized moments
of the ambient alternative distribution}
}
\description{
Creates a `boltzmannTestResult` object
}
