% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\name{dplyr_row_slice.entities_tibble}
\alias{dplyr_row_slice.entities_tibble}
\title{`dplyr` row slicing}
\usage{
\method{dplyr_row_slice}{entities_tibble}(object, i, ...)
}
\arguments{
\item{object}{an `entities_tibble` object}

\item{i}{row indices selected by an upstream `dplyr` function}

\item{...}{additional parameters}
}
\description{
This function is automatically called after the `dplyr` functions `filter`, `slice`,
`arrange` etc. It automatically updates (marginalize) the empirical relative frequencies
and the sample size
}
