% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes_tibble.R
\name{dplyr_row_slice.outcomes_tibble}
\alias{dplyr_row_slice.outcomes_tibble}
\title{`dplyr` row slicing for an `outcomes_tibble`}
\usage{
\method{dplyr_row_slice}{outcomes_tibble}(object, i, ...)
}
\arguments{
\item{object}{an `outcomes_tibble` object}

\item{i}{row indices selected by an upstream `dplyr` function}

\item{...}{additional parameters}
}
\value{
A sliced `outcomes_tibble`
}
\description{
This function is automatically called after the `dplyr` functions `filter`,
`slice`, `arrange` etc. It automatically updates (marginalize) the empirical
relative frequencies and the sample size
}
\examples{
data(kidneyStones)
outcomes <- outcomes_tibble(kidneyStones)

## remove small stones
(outcomesLargeStones <- dplyr::filter(outcomes, stoneSize != "small"))

## this has adapted the empirical attribute to reflect the smaller
## sample size (from 700 to 343)
sum(empirical(outcomesLargeStones))


}
