% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes_tibble.R
\name{empirical<-}
\alias{empirical<-}
\title{Set the empirical distribution}
\usage{
empirical(object, tolerance = .Machine$double.eps) <- value
}
\arguments{
\item{object}{an `outcomes_tibble` object}

\item{value}{an non-negative numeric vector of the same length
as the empirical vector that sums to 1. Corresponds to the new empirical
distribution,}
}
\description{
Setter for the `empirical` attribute of an `outcomes_tibble` object
}
