% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\name{entities_tibble}
\alias{entities_tibble}
\title{Build an `entities_tibble` class object}
\usage{
entities_tibble(data)
}
\arguments{
\item{data}{a tibble-like object}
}
\description{
`entities_tibble()` constructs a subclass of `tibble` with unique rows from an input tibble-like object. It adds two attributes
\itemize{
  \item{`empirical`}{the empirical relative frequencies of the entities (unique rows)}
  \item{`sampleSize`}{the sample size given by the number of rows of the input `data.frame`}
}
}
