% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eta_isFeasible}
\alias{eta_isFeasible}
\title{Check whether a vector is a feasible linear combination on the simplex}
\usage{
eta_isFeasible(G, eta, tol = .Machine$double.eps)
}
\arguments{
\item{G}{a numeric matrix}

\item{eta}{a numeric vector with one value per row of G}

\item{tol}{double (default `.Machine$double.eps`) tolerance}
}
\value{
bool with `TRUE` if G p = eta is feasible
}
\description{
Checks whether a vector `eta` = `G` `p` can be obtained by a vector `p` with
p_i >= 0
and sum p_i = 1
}
\keyword{internal}
