% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProjector.R
\name{iProjector}
\alias{iProjector}
\title{Compute the I-projection}
\usage{
iProjector(G, eta, v, maxit = 10000L, convTolerance = .Machine$double.eps)
}
\arguments{
\item{G}{matrix with `nrow(G)` constraints (including normalization) for `ncol(G)` entities}

\item{eta}{vector with one value for each row of `G`}

\item{v}{reference distribution with one value for each column of `G`}

\item{maxit}{single integer value (default = 10000L) for the maximimal number of iterations}

\item{convTolerance}{double value (default = `.Machine$double.eps` for the convergence tolerance}
}
\description{
Computes the I-projection of v onto the linear family G p = eta. Throws a warning if any condition cannot be met and returns the
reference distribution `v`.
}
