% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix_hasFullRowRank}
\alias{matrix_hasFullRowRank}
\title{Check whether a matrix has full row rank}
\usage{
matrix_hasFullRowRank(G, tol = .Machine$double.eps)
}
\arguments{
\item{G}{a numeric matrix with #rows <= #columns}

\item{tol}{double (default `.Machine$double.eps`) tolerance}
}
\value{
bool with `TRUE` if G has full row rank and `FALSE` if not
}
\description{
Uses the QR decomposition to check whether a matrix `G` has full row rank
}
\keyword{internal}
