% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomialPvalue.R
\name{multinomialPvalue}
\alias{multinomialPvalue}
\title{Compute exact multinomial p-values}
\usage{
multinomialPvalue(G, eta, mu, v, N)
}
\arguments{
\item{G}{matrix with `nrow(G)` constraints (does not require normalization)
for `ncol(G)` entities}

\item{eta}{vector with one value for each row of `G`}

\item{mu}{vector with the empirical values}

\item{v}{hypothesis distribution with one value for each column of `G`}

\item{N}{samples size}
}
\value{
the exact multinomial p-value
}
\description{
Computes the exact multinomial p-value using
}
\examples{

data <- data.frame(
  x = c(rep(-1, 4), rep(0, 4), rep(1, 2))
)
(bt <- boltzmann.test(~x, data = data))

## exact multinomial p-value
(exactPvalue <- multinomialPvalue(
  G = bt$coefficientMatrix[-1, , drop = FALSE],
  eta = bt$hypothesisExpectations[-1],
  mu = bt$alternativeExpectations[-1],
  v = bt$hypothesisDistribution,
  N = bt$sampleSize
))


}
