% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\name{new_entities_tibble}
\alias{new_entities_tibble}
\alias{validate_entities_tibble}
\title{Constructor and validator for an `entities_tibble` class object}
\usage{
new_entities_tibble(data, empirical, sampleSize)

validate_entities_tibble(object)
}
\arguments{
\item{data}{A tibble-like object with unique rows}

\item{empirical}{a vector of empirical relative frequencies with one value per row of `data`}

\item{sampleSize}{a single numeric value giving the sample size. Should be at least the number of
rows in `data`.}

\item{object}{an entities_tibble class object}
}
\description{
Creates an `entities_tibble` class object. From a `tibble` with unique rows (=entities) and
the empirical relative frequencies of the entities in the input `data.frame` and the sample size
\itemize{
  \item{`empirical`}{the empirical relative frequencies of the entities (unique rows)}
  \item{`sampleSize`}{the sample size given by the number of rows of the input `data.frame`}
}

`validate_entities_tibble()` checks an `entities_tibble` for internal consistency.
}
