% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes_tibble.R
\name{new_outcomes_tibble}
\alias{new_outcomes_tibble}
\alias{validate_outcomes_tibble}
\title{Constructor and validator for an `outcomes_tibble` class object}
\usage{
new_outcomes_tibble(data, empirical, sampleSize)

validate_outcomes_tibble(object)
}
\arguments{
\item{data}{A tibble-like object with unique rows}

\item{empirical}{a vector of empirical relative frequencies with one value per row of `data`}

\item{sampleSize}{a single numeric value giving the sample size. Should be at least the number of
rows in `data`.}

\item{object}{an outcomes_tibble class object}
}
\description{
Creates an `outcomes_tibble` class object from a data frame. The `outcomes_tibble` will have unique rows (=outcomes) and
the empirical relative frequencies of the outcomes in the input `data.frame` and the sample size
\itemize{
  \item{`empirical`}{the empirical relative frequencies of the outcomes (unique rows)}
  \item{`sampleSize`}{the sample size given by the number of rows of the input `data.frame`}
}

`validate_outcomes_tibble()` checks an `outcomes_tibble` for internal consistency.
}
