% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\name{print.entities_tibble}
\alias{print.entities_tibble}
\title{Print Method for class `entities_tibble`}
\usage{
\method{print}{entities_tibble}(object, ...)
}
\arguments{
\item{object}{an `entities_tibble` object}

\item{...}{additional arguments for print}
}
\value{
Function prints to console
}
\description{
Appends a column empirical and prints the sample size
}
