% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes_tibble.R
\name{print.outcomes_tibble}
\alias{print.outcomes_tibble}
\title{Print Method for class `outcomes_tibble`}
\usage{
\method{print}{outcomes_tibble}(object, ...)
}
\arguments{
\item{object}{an `outcomes_tibble` object}

\item{...}{additional arguments for print}
}
\value{
Function prints to console
}
\description{
Appends a column empirical and prints the sample size
}
