% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\name{sampleSize}
\alias{sampleSize}
\title{Obtain the sample size}
\usage{
sampleSize(object)
}
\arguments{
\item{object}{an `entities_tibble` object}
}
\value{
the sample size
}
\description{
Accessor for the `sampleSize` attribute of an `entities_tibble` object
}
