% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\name{[.entities_tibble}
\alias{[.entities_tibble}
\title{Subset an `entities_tibble`}
\usage{
\method{[}{entities_tibble}(object, i, j, drop = FALSE)
}
\arguments{
\item{object}{an `entities_tibble` object}

\item{i}{row indices or logical vector of length `nrow(object)`}

\item{j}{col indices, names, or logical vector of length `ncol(object)`}

\item{drop}{logical (default = FALSE). If `TRUE`, simplifies the result if possible.}
}
\value{
An `entities_tibble` object with updated `empirical` and `sampleSize` attributes.
#' @export
}
\description{
Subsetting an `entities_tibble` works similarily to subsetting a `tibble`,
but the object maintains an aligned `empirical` frequency attribute and a correspond
sample size
}
\details{
- When subsetting rows, the relative frequencies (`empirical`) are **renormalized**
  to sum to 1 for the selected rows.
- The `sampleSize` attribute is updated to reflect the sum of the original
  data points represented by the selected rows.
- Column subsetting (`j`) works like a normal tibble; attributes are preserved.
}
