% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes_tibble.R
\name{[.outcomes_tibble}
\alias{[.outcomes_tibble}
\title{Subset an `outcomes_tibble`}
\usage{
\method{[}{outcomes_tibble}(object, i, j, drop = FALSE)
}
\arguments{
\item{object}{an `outcomes_tibble` object}

\item{i}{row indices or logical vector of length `nrow(object)`}

\item{j}{col indices, names, or logical vector of length `ncol(object)`}

\item{drop}{logical (default = FALSE). If `TRUE`, simplifies the result if
possible.}
}
\value{
An `outcomes_tibble` object with updated `empirical` and
`sampleSize` attributes.
}
\description{
Subsetting an `outcomes_tibble` works similarily to subsetting a `tibble`,
but the object maintains an aligned `empirical` frequency attribute and
sample size
}
\details{
- When subsetting rows, the relative frequencies (`empirical`) are
  **renormalized**
  to sum to 1 for the selected rows.
- The `sampleSize` attribute is updated to reflect the sum of the original
  data points represented by the selected rows.
- Column subsetting (`j`) works like a normal tibble; attributes are
  preserved.
}
