% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boltzmannTestResult.R
\name{validate_boltzmannTestResult}
\alias{validate_boltzmannTestResult}
\title{Validate a `boltzmannTestResult` object}
\usage{
validate_boltzmannTestResult(object)
}
\arguments{
\item{object}{a `boltzmannTestResult` class object}
}
\description{
Checks a `boltzmannTestResult` object for internal consistency.
}
\keyword{internal}
