% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes_tibble.R
\name{validate_outcomes_tibble}
\alias{validate_outcomes_tibble}
\title{Validate an `outcomes_tibble` class object}
\usage{
validate_outcomes_tibble(object)
}
\arguments{
\item{object}{an outcomes_tibble class object}
}
\value{
An `outcomes_tibble` object. If validation fails it stops with
an error message about what could not be validated.
}
\description{
Checks an `outcomes_tibble` for internal consistency.
}
\keyword{internal}
