% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which_approx_equal}
\alias{which_approx_equal}
\title{Check which entries of two numeric vectors are (approximately) equal}
\usage{
which_approx_equal(a, b, tol = .Machine$double.eps, reldiff = FALSE)
}
\arguments{
\item{a}{a numeric vector}

\item{b}{a numeric vector of the same length as a}

\item{tol}{double value (default`.Machine$double.eps`) of the tolerance}

\item{reldiff}{logical (default `FALSE`) whether to compute the relative
difference `TRUE` or the absolute difference `FALSE`}
}
\value{
a logical vector with elements `TRUE` if the entries are a
pproximately equal or `FALSE` if not.
}
\description{
Checks which entries of two numeric vector are up to a difference of > `tol`
equal. Note that we follow the convention of Armadillo: if the difference of
two numbers x and y is smaller or equal to `tol` then the numbers are
approximately equal.
}
\keyword{internal}
