% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boltzmann.test.R
\name{boltzmann.test}
\alias{boltzmann.test}
\alias{boltzmann.test.outcomes_tibble}
\alias{boltzmann.test.numeric}
\alias{boltzmann.test.formula}
\title{Perform a Boltzmann Test}
\usage{
boltzmann.test(object, ...)

\method{boltzmann.test}{outcomes_tibble}(
  outcomes,
  G,
  eta,
  testedExpectations,
  nestedExpectations = NULL,
  maxit = 10000L,
  tolerance = .Machine$double.eps
)

\method{boltzmann.test}{numeric}(x, y = NULL, mu = 0, paired = FALSE)

\method{boltzmann.test}{formula}(formula, data, nu = 0)
}
\arguments{
\item{outcomes}{an `outcomes_tibble` object}

\item{G}{a numeric matrix with the function values for the outcomes in the rows
and the outcomes in the columns}

\item{eta}{a numeric vector with the same number of entries as `G` has rows}

\item{testedExpectations}{an integer vector indicating the tested expectations}

\item{nestedExpectations}{an integer vector indicating the nested expectations}

\item{maxit}{an integer specifying the maximimal number of iterations used in
the I-projector (default 10000L)}

\item{tolerance}{a numeric specifying the numeric tolerance (default
.Machine$double.eps)}

\item{x}{a (non-empty) numeric vector of data values}

\item{y}{an optional (non-empty) numeric vector of data values}

\item{mu}{a single number indicating the true value of the mean or the difference in
means when performing a two sample test}

\item{paired}{a logical indicating whether to perform a paired test (default
FALSE)}

\item{formula}{a formula either of the form `lhs` ~ `rhs`, where `lhs` is either
a numeric or factor variable or combinations thereof (see details) and
`rhs` is one or more factor variable used to define the groups, or of the form
~ `rhs` where `rhs` contains one or more numeric or factor variable (see details)}

\item{data}{a data.frame containing the variables used in `formula`}

\item{nu}{a single number or a numeric vector specifying the values of the
expectations}
}
\description{
`boltzmann.test` with `object` being an outcome_tibble is the general
interface to perform the Boltzmann Test.

`boltzmann.test` with `object` being a numeric vector performs one or two sample
Boltzmann Tests for the mean. It follows the `t.test` interface.

`boltzmann.test` with `object` being a formula performs multigroup
or multivariate comparisons
}
\details{
The `formula` API allows to specify multigroup comparisons for one or more target
variable specified at `lhs`. If a target variable is numeric then the difference
of means of all groups except the first versus the mean of the first group is compared.
if a target variable is a factor then the difference of the prevalence of the second (third, etc) level
of that level of all groups except the first versus the first group is compared.
If only a `rhs` is given, a multivariate comparison to the given `nu` is performed. If the target
is numeric the mean is calculated. If the target is a factor the prevalence is calculated.
}
