% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boltzmannTest-package.R
\docType{package}
\name{boltzmannTest-package}
\alias{boltzmannTest}
\alias{boltzmannTest-package}
\title{boltzmannTest: Non-parametric hypothesis testing with the Boltzmann Test}
\description{
Perform non-parametric hypothesis testing about expectations with the Boltzmann Test.
}
\author{
\strong{Maintainer}: Ho-Ryun Chung \email{ho.chung@uni-marburg.de}

Authors:
\itemize{
  \item Orestis Loukas
}

}
\keyword{internal}
