% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boltzmannTestResult.R
\name{boltzmannTestResult}
\alias{boltzmannTestResult}
\title{Build a `boltzmannTestResult` class object}
\usage{
boltzmannTestResult(
  statistic,
  iDivergence,
  degreesOfFreedom,
  sampleSize,
  pValue,
  alternativeDistribution,
  alternativeExpectations,
  hypothesisDistribution,
  hypothesisExpectations,
  testedExpectations,
  dataName,
  coefficientMatrix,
  method = "Boltzmann Test",
  ambientDistribution = NULL,
  ambientExpectations = NULL
)
}
\arguments{
\item{statistic}{a single numeric value corresponding to the value of the
(test) statistic}

\item{iDivergence}{a single numeric value corresponding to the value of the
I-Divergence of the hypothesis distribution or outer alternative
distribution to the (inner) alternative distribution}

\item{degreesOfFreedom}{a single integer value for the degrees of freedom
of the Chi^2 distribution. Gives the number of tested generalized
Expectations and must be > 0}

\item{sampleSize}{a single integer value for the sample size. Must be > 1}

\item{pValue}{a single numeric value between 0 and 1 for the p-value}

\item{alternativeDistribution}{a numeric vector for the alternative
distribution (usually equal to the empirical distribution)}

\item{alternativeExpectations}{a numeric vector for the values of the
observed expectations of the alternative distribution}

\item{hypothesisDistribution}{a numeric vector for the hypothesis
distribution}

\item{hypothesisExpectations}{a numeric vector for the values of the
expectations of the hypothesis distribution}

\item{testedExpectations}{integer vector with the indices of the
tested expectations}

\item{dataName}{string with the name of the data}

\item{coefficientMatrix}{a numeric matrix holding the coefficient matrix
to calculate the expectations. The number of rows must equal the number of
expectations and the number of columns the number of entities.}

\item{ambientDistribution}{a numeric vector for the ambient alternative
distribution}

\item{ambientExpectations}{a numeric vector for the values of the
expectations of the ambient alternative distribution}
}
\value{
a list with class `boltzmannTestResult` containing the follow
components
\describe{
 \item{statistic}{the value of \eqn{\chi^2} statistic}
 \item{iDivergence}{the value of the I-divergence, a.k.a. Kullback-Leibler
 divergence of the hypothesis distribution or ambient alternative
 distribution to the empirical distribution}
 \item{degreesOfFreedom}{ the degrees of freedom used for computing the
 p-value via the \eqn{\chi^2} distribution with `degreesOfFreedom`}
 \item{sampleSize}{the sample size}
 \item{pValue}{the p-value}
 \item{alternativeDistribution}{the empirical distribution of outcomes}
 \item{alternativeExpectations}{the values of empirical expectations}
 \item{hypothesisDistribution}{the hypothesis distribution of outcomes}
 \item{hypothesisExpectations}{the values of hypothesized expectations}
 \item{testedExpectations}{an integer vector with the indices of the tested
   expectations}
 \item{dataName}{name of the input data}
 \item{coefficientMatrix}{an numeric matrix with the coefficient matrix}
 \item{method}{a string signifying the method}
 \item{ambientDistribution}{the ambient alternative distribution of outcomes
   if nested hypothesis testing is performed otherwise NULL}
 \item{ambientExpectations}{the values of the ambient expectations if nested
   hypothesis testing is performed otherwise NULL}
}
}
\description{
Creates a `boltzmannTestResult` object
}
