% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes_tibble.R
\name{empirical}
\alias{empirical}
\alias{empirical<-}
\title{Get or set the `empirical` attribute of an `outcomes_tibble` object}
\usage{
empirical(object)

empirical(object) <- value
}
\arguments{
\item{object}{an `outcomes_tibble` object}

\item{value}{an non-negative numeric vector of the same length
as the empirical vector that sums to 1. Corresponds to the new empirical
distribution,}
}
\value{
a numeric vector of length = `nrow(object)` with the empirical
relative frequencies
}
\description{
`empirical()` returns the `empirical` attribute of an `outcomes_tibble`
object, i.e. the empirical frequencies of the outcomes

`empirical<-` sets the `empirical` attribute of an `outcomes_tibble` object
}
\examples{
data(nhanes)
outcomes <- outcomes_tibble(nhanes)

## get the empirical distribution
f <- empirical(outcomes)


data(nhanes)
outcomes <- outcomes_tibble(nhanes)

## set the empirical distribution to the MEC2 sample weights
empirical(outcomes) <- outcomes$WTMEC2YR / sum(outcomes$WTMEC2YR)

}
