% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes_tibble.R
\name{empirical}
\alias{empirical}
\title{Obtain the empirical relative frequencies}
\usage{
empirical(object)
}
\arguments{
\item{object}{an `outcomes_tibble` object}
}
\value{
a numeric vector of length = `nrow(object)` with the empirical relative
frequencies
}
\description{
Accessor for the `empirical` attribute of an `outcomes_tibble` object
}
