% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iDivergence.R
\name{iDivergence}
\alias{iDivergence}
\title{Computes the I-divergence}
\usage{
iDivergence(p, q, tolerance = .Machine$double.eps)
}
\arguments{
\item{p}{a numeric vector that should sum up to one (candidate distribution)}

\item{q}{a numeric vector that should sum up to one (reference distribution)}

\item{tolerance}{the numerical tolerance (default .Machine$double.eps)}
}
\value{
the value of the I-divergence (a.k.a. Kullback-Leibler divergence)
of q to p
}
\description{
Given two probability vectors p and q compute the I-divergence
also known as Kullback-Leibler divergence D(p||q) of p from q
}
\examples{

data(kidneyStones)
outcomes <- outcomes_tibble(kidneyStones)
G <- with(
  outcomes,
  rbind(
    norm = 1,
    # structural expectation fraction treatment A
    treatment_A = treatment == "A",
    # ambient expectation fraction small stone size given treatment B
    stoneSize_small.treatment_B =
      (stoneSize == "small" & treatment == "B") / 0.5,
    # ambient expectation fraction small stone size given treatment A
    stoneSize_small.treatment_A =
      (stoneSize == "small" & treatment == "A") / 0.5
  )
)
eta <- c(
  norm = 1,
  treatment_A = 0.5,
  stoneSize_small.treatment_B = 0.51,
  stoneSize_small.treatment_A = 0.51

)
## perform I-projection
iProj <- iProjector(G, eta = eta, v = empirical(outcomes))

iDivergence(iProj$p, empirical(outcomes))


}
