% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iDivergence.R
\name{iDivergence}
\alias{iDivergence}
\title{Computes the I-divergence}
\usage{
iDivergence(p, q, tolerance = .Machine$double.eps)
}
\arguments{
\item{p}{a numeric vector that should sum up to one (candidate distribution)}

\item{q}{a numeric vector that should sum up to one (reference distribution)}
}
\value{
the value of the I-divergence (a.k.a. Kullback-Leibler divergence)
of q to p
}
\description{
Given two probability vectors p and q compute the I-divergence
also known as Kullback-Leibler divergence D(p||q) of p from q
}
