% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProjector.R
\name{iProjector}
\alias{iProjector}
\title{Compute the I-projection}
\usage{
iProjector(G, eta, v, maxit = 10000L, convTolerance = .Machine$double.eps)
}
\arguments{
\item{G}{matrix with `nrow(G)` constraints (including normalization) for
`ncol(G)` entities}

\item{eta}{vector with one value for each row of `G`}

\item{v}{reference distribution with one value for each column of `G`}

\item{maxit}{single integer value (default = 10000L) for the maximimal number
of iterations}

\item{convTolerance}{double value (default = `.Machine$double.eps` for the
convergence tolerance}
}
\value{
A list with the following entries
\describe{
 \item{p}{the I-projection of v onto the linear family described by
 G p = eta}
 \item{converged}{a integer value, with 0 no convergence, 1 converged &
 conditions are met and 2 converged & conditions are not met}
 \item{iter}{the number of iterations used}
 \item{error}{an error message if any}
}
}
\description{
Computes the I-projection of v onto the linear family G p = eta. If the
conditions cannot be met, the function throws a warning and returns the
reference distribution v
}
\examples{

## projecting the kidney stone data to a study, where the
## treatments A and B received the same proportion of small and large stones
data(kidneyStones)
outcomes <- outcomes_tibble(kidneyStones)
G <- with(
  outcomes,
  rbind(
    norm = 1,
    # structural expectation fraction treatment A
    treatment_A = treatment == "A",
    # ambient expectation fraction small stone size given treatment B
    stoneSize_small.treatment_B =
      (stoneSize == "small" & treatment == "B") / 0.5,
    # ambient expectation fraction small stone size given treatment A
    stoneSize_small.treatment_A =
      (stoneSize == "small" & treatment == "A") / 0.5
  )
)
eta <- c(
  norm = 1,
  treatment_A = 0.5,
  stoneSize_small.treatment_B = 0.51,
  stoneSize_small.treatment_A = 0.51

)
## perform I-projection
iProj <- iProjector(G, eta = eta, v = empirical(outcomes))

## the old difference in the treatment success rate was
with(
  outcomes,
  sum(
    empirical(outcomes) * (treatment == "B") * (success == "yes")
  ) / sum(empirical(outcomes) * (treatment == "B")) -
  sum(
    empirical(outcomes) * (treatment == "A") * (success == "yes")
  ) / sum(empirical(outcomes) * (treatment == "A"))
)

## the "corrected" difference in the treatment success rate is
with(
  outcomes,
  sum(
    iProj$p * (treatment == "B") * (success == "yes")
  ) / sum(iProj$p * (treatment == "B")) -
  sum(
    iProj$p * (treatment == "A") * (success == "yes")
  ) / sum(iProj$p * (treatment == "A"))
)

}
