% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomialPvalue.R
\name{multinomialPvalue}
\alias{multinomialPvalue}
\title{Compute exact multinomial p-values}
\usage{
multinomialPvalue(G, eta, mu, v, N)
}
\arguments{
\item{G}{matrix with `nrow(G)` constraints (does not require normalization) for `ncol(G)` entities}

\item{eta}{vector with one value for each row of `G`}

\item{mu}{vector with the empirical values}

\item{v}{reference distribution with one value for each column of `G`}

\item{N}{samples size}
}
\description{
Computes the exact multinomial p-value using
}
