% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes_tibble.R
\name{new_outcomes_tibble}
\alias{new_outcomes_tibble}
\title{Construct for an `outcomes_tibble` object}
\usage{
new_outcomes_tibble(data, empirical, sampleSize)
}
\arguments{
\item{data}{a data frame with outcomes in the rows}

\item{empirical}{a numeric vector with one value per row of `data`}

\item{sampleSize}{an integer value indicating the sample Size}
}
\value{
An `outcomes_tibble` object
}
\description{
Constructs an `outcomes_tibble`  object with unique
rows from an input tibble-like object. It adds two attributes
\itemize{
  \item{`empirical` the empirical relative frequencies of the outcomes
  (unique rows)}
  \item{`sampleSize` the sample size given by the number of rows of the
  input `data.frame`}
}
}
\keyword{internal}
