% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes_tibble.R
\name{outcomes_tibble}
\alias{outcomes_tibble}
\title{Build an `outcomes_tibble` class object}
\usage{
outcomes_tibble(data)
}
\arguments{
\item{data}{a data frame}
}
\description{
Constructs an `outcomes_tibble()`  a subclass of `tibble` with unique rows from an input tibble-like object. It adds two attributes
\itemize{
  \item{`empirical` the empirical relative frequencies of the outcomes (unique rows)}
  \item{`sampleSize` the sample size given by the number of rows of the input `data.frame`}
}
}
