% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes_tibble.R
\name{sampleSize}
\alias{sampleSize}
\title{Get the sample size}
\usage{
sampleSize(object)
}
\arguments{
\item{object}{an `outcomes_tibble` object}
}
\value{
the sample size
}
\description{
Accessor for the `sampleSize` attribute of an `outcomes_tibble` object
}
\examples{
data(nhanes)
outcomes <- outcomes_tibble(nhanes)

sampleSize <- sampleSize(outcomes)

}
